# Improve performance of NewTensor and Value for string tensors